refactor(routes): move mjpg endpoints to blueprint
